refactor(apple): Use string for URL configuration (#9124)
In the UI, we need to use Strings to bind to the text inputs.
In the configuration dictionaries, we need to use Strings to save the
URLs.

It makes no sense to convert these to URLs in between. Instead, we can
validate upon save and then use them as Strings throughout.